Embedded Voltmeter Initial Code Commit
It compiles, it connects to networks and will connect to the server. Now
I'm working on getting it to send useful messages. Soon I'll port in
some of the code from device.cpp and client.cpp, not all of it due to
memory limitations but some. Loook at wlan.h for some fun variables to
play with.