Updated App With our functioning ADC10!!!
ADC10 code has been added to the basicwifi.c along with all of its
functions and pragma interrupts.

The out puts of the ADC10 process are global variables called

ADC10_READING and battLow they are also located in basicwifi.c right
above the ADC10Routine function prototype just above main.

right now the adc10 timer is set for 1 minute between readings this can
be easily changed if needed, also the funtionality of ADC10Routine is
currently implementing our waiting period before transmissions as some
of the clocks of interest are used by external sources in the while(1)
loop of main() - I figured it would be easier to implement the wait
period here and restore clock settings before exiting the ADC10 function
so we do not risk messing with the rest of the code.

Cheers guys