Time stamps fixed, sends readings.
Have to make sure that the math is using longs, and so have to bitwise
AND with 0xffffffffL note the L. The L mean long so it'll do shifts
greater than 15 bits. Also fixed the time code and now it will send
readings every 30 seconds. Next have to add frequency sending from
server and incorporate that. Also reduced the message size to 8 bits
(from 12), 4 for time and 4 for data.